chore(dtl-lexer): move tag related rust files to tag module
